CAMEL-19807: Move spring XML <beans> logic to its own class
